requirements.txt
addict
nltk
PyYAML
names
inflect
networkx
pandas
tqdm
comet_ml
torch
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
matplotlib
seaborn
bert-serving-client
allennlp